+++ dali-toolkit/internal/visuals/wireframe/wireframe-visual.cpp
@@ -43,12 +43,12 @@ const char * const INDEX_NAME("indices");
 const char* VERTEX_SHADER = DALI_COMPOSE_SHADER(
 attribute mediump vec2  aPosition;\n
 uniform   highp   mat4  uMvpMatrix;\n
-uniform   mediump vec3  uSize;\n
+uniform   highp   vec3  uSize;\n
 \n
 
 //Visual size and offset
 uniform mediump vec2 offset;\n
-uniform mediump vec2 size;\n
+uniform highp   vec2 size;\n
 uniform mediump vec4 offsetSizeMode;\n
 uniform mediump vec2 origin;\n
 uniform mediump vec2 anchorPoint;\n
